Fix import issues with validators
`checkUsernameRequirements` and `checkPasswordRequirements` have been separated into their own files.
